refactor: [GSW-2274] sortDisplayRewards
